+++ packages/platform-android/src/index.ts
@@ -55,13 +55,6 @@ export async function readAndroidAppState(
   return await read(host, device, signal);
 }
 
-export async function parseAndroidForegroundApp(
-  text: string,
-): Promise<Readonly<{ package?: string; activity?: string }> | null> {
-  const { parseAndroidForegroundApp: parse } = await import('./app-state.ts');
-  return parse(text);
-}
-
 export const runtimeModule = Object.freeze({
   ...metadata,
   loadRuntime: async (host) => {
