+++ scripts/layering/package-boundaries.test.ts
@@ -75,6 +75,7 @@ const CONTRACT_EXPORTS = [
   '@agent-device/contracts/audio-probe-support',
   '@agent-device/contracts/audio-runtime-plan',
   '@agent-device/contracts/back-mode',
+  '@agent-device/contracts/backend-diagnostics',
   '@agent-device/contracts/back-runtime',
   '@agent-device/contracts/boot-failure',
   '@agent-device/contracts/capture',
@@ -148,6 +149,7 @@ const CONTRACT_EXPORTS = [
   '@agent-device/contracts/settings',
   '@agent-device/contracts/settings-runtime',
   '@agent-device/contracts/snapshot',
+  '@agent-device/contracts/snapshot-desktop-projection',
   '@agent-device/contracts/snapshot-presentation',
   '@agent-device/contracts/snapshot-runtime',
   '@agent-device/contracts/snapshot-timeout-evidence',
