+++ src/__tests__/eager-closure-budgets.ts
@@ -172,6 +172,7 @@ export const FACADE_BUDGETS: Readonly<Record<string, number>> = Object.freeze({
   'packages/contracts/src/audio-probe-support.ts': 5,
   'packages/contracts/src/audio-runtime-plan.ts': 5,
   'packages/contracts/src/back-mode.ts': 1,
+  'packages/contracts/src/backend-diagnostics.ts': 1,
   'packages/contracts/src/boot-failure.ts': 1,
   'packages/contracts/src/click-button.ts': 3,
   'packages/contracts/src/clipboard.ts': 1,
@@ -237,6 +238,7 @@ export const FACADE_BUDGETS: Readonly<Record<string, number>> = Object.freeze({
   'packages/contracts/src/scroll-runtime.ts': 4,
   'packages/contracts/src/selector-observation-runtime.ts': 1,
   'packages/contracts/src/settings.ts': 3,
+  'packages/contracts/src/snapshot-desktop-projection.ts': 2,
   'packages/contracts/src/snapshot-presentation.ts': 2,
   'packages/contracts/src/snapshot-runtime.ts': 3,
   'packages/contracts/src/snapshot-timeout-evidence.ts': 1,
